Added fire sounds
