+++ Game Audio FMOD/Metadata/AudioFile/{9596221e-d8ae-4509-a7f8-ecea40cf9277}.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<objects serializationModel="Studio.02.02.00">
+	<object class="AudioFile" id="{9596221e-d8ae-4509-a7f8-ecea40cf9277}">
+		<property name="assetPath">
+			<value>Fire/Fire-Crackle-1.wav</value>
+		</property>
+		<property name="frequencyInKHz">
+			<value>48</value>
+		</property>
+		<property name="channelCount">
+			<value>2</value>
+		</property>
+		<property name="length">
+			<value>2.4428749999999999</value>
+		</property>
+		<relationship name="masterAssetFolder">
+			<destination>{8354e7ce-d72e-419f-a828-b62749a1431b}</destination>
+		</relationship>
+	</object>
+</objects>
