+++ Game Audio FMOD/Metadata/Event/{33644cd5-2352-4928-af9d-1d660fcd9ff2}.xml
@@ -63,7 +63,7 @@
 			<destination>{268d83b0-267e-4add-97db-dbb94966d987}</destination>
 		</relationship>
 		<relationship name="output">
-			<destination>{b9b09b37-179e-4ee6-a53e-12791ec3edd6}</destination>
+			<destination>{f036f4e1-a7bc-4dd3-b7bd-317f4bb2373d}</destination>
 		</relationship>
 	</object>
 	<object class="EventAutomatableProperties" id="{9ab06e16-6051-4b90-bf2d-36c8e30cc905}" />
