+++ Game Audio FMOD/Metadata/EventFolder/{a09a7d4e-bcb5-4051-bddd-2fa536ab4f84}.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<objects serializationModel="Studio.02.02.00">
+	<object class="EventFolder" id="{a09a7d4e-bcb5-4051-bddd-2fa536ab4f84}">
+		<property name="name">
+			<value>Fire</value>
+		</property>
+		<relationship name="folder">
+			<destination>{59ebdec8-aff5-4e35-b5dd-dbc8844f3d3e}</destination>
+		</relationship>
+	</object>
+</objects>
