+++ Game Audio/Assets/Work/Scripts/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d6513b82d1e4e6a48bff177e297a3180
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
