+++ Game Audio/Assets/Work/Scripts/Sounds/Torch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff6639572fdcc904f9cf11a0acf169d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
