+++ Game Audio/Assets/Work/Scripts/Sounds/Wind.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad3c4e9e2d3203f42b2f35a4aa66f130
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
